+++ src/main/kotlin/me/rafaelldi/einburgerungstest/questions/QuestionCategories.kt
@@ -1,0 +1,7 @@
+package me.rafaelldi.einburgerungstest.questions
+
+internal data class QuestionCategories(
+    val groupCategories: List<Pair<QuestionCategory, Int>> = emptyList(),
+    val nationalCategories: List<Pair<QuestionCategory, Int>> = emptyList(),
+    val regionalCategories: List<Pair<QuestionCategory, Int>> = emptyList()
+)