+++ src/main/kotlin/me/rafaelldi/einburgerungstest/toolWindow/EinburgerungstestViewModel.kt
@@ -7,14 +7,17 @@ import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.flow.updateAndGet
 import me.rafaelldi.einburgerungstest.persistence.QuestionPersistenceService
 import me.rafaelldi.einburgerungstest.questions.Question
+import me.rafaelldi.einburgerungstest.questions.QuestionCategories
 import me.rafaelldi.einburgerungstest.questions.QuestionCategory
 import me.rafaelldi.einburgerungstest.questions.QuestionQuizService
 
 internal interface EinburgerungstestViewModel : Disposable {
     val uiState: StateFlow<UiState>
+    val questionCategories: StateFlow<QuestionCategories>
     val selectedCategory: StateFlow<QuestionCategory>
     val currentQuestion: StateFlow<Pair<Question, ImageBitmap?>?>
     val selectedAnswerIndex: StateFlow<Int?>
@@ -41,6 +44,9 @@ internal class EinburgerungstestViewModelImpl(
     private val _uiState = MutableStateFlow<UiState>(UiState.Loading)
     override val uiState: StateFlow<UiState> = _uiState.asStateFlow()
 
+    private val _questionCategories = MutableStateFlow(QuestionCategories())
+    override val questionCategories: StateFlow<QuestionCategories> = _questionCategories.asStateFlow()
+
     private val _selectedCategory = MutableStateFlow(persistence.selectedCategory)
     override val selectedCategory: StateFlow<QuestionCategory> = _selectedCategory.asStateFlow()
 
@@ -58,7 +64,18 @@ internal class EinburgerungstestViewModelImpl(
 
     override suspend fun loadQuestions() {
         quizService.loadQuestions()
+        updateQuestionCategories()
         _uiState.value = UiState.NotStarted
+    }
+
+    private fun updateQuestionCategories() {
+        _questionCategories.update {
+            QuestionCategories(
+                groupCategories = QuestionCategory.groupCategories.map { it to quizService.getQuestionCount(it) },
+                nationalCategories = QuestionCategory.nationalCategories.map { it to quizService.getQuestionCount(it) },
+                regionalCategories = QuestionCategory.regionalCategories.map { it to quizService.getQuestionCount(it) }
+            )
+        }
     }
 
     override fun onStartQuiz() {
@@ -112,6 +129,7 @@ internal class EinburgerungstestViewModelImpl(
             updatedList
         }
         persistence.favorites = list
+        updateQuestionCategories()
     }
 
     override fun dispose() {